Add eslint-plugin-flowtype
