+++ packages/benchmarks/src/app/ReportCard.js
@@ -1,3 +1,5 @@
+/* @noflow */
+
 import Text from './Text';
 import { StyleSheet, View } from 'react-native';
 import React, { Fragment } from 'react';
