+++ packages/benchmarks/src/impl.js
@@ -1,3 +1,5 @@
+/* @noflow */
+
 import { type Component } from 'react';
 import packageJson from '../package.json';
