+++ packages/docs/src/components/FlatList/examples/MultiColumn.js
@@ -4,7 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @providesModule MultiColumnExample
+ * @noflow
  */
 'use strict';
