+++ packages/docs/src/components/FlatList/examples/SingleColumn.js
@@ -3,6 +3,8 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
+ *
+ * @noflow
  */
 'use strict';
